src/main/java/survey/dto/Survey.java
package survey.dto;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Table;

import org.springframework.data.annotation.Id;

/**
 * Entity representing customers feedback
 *
 * @author felipey.
 */
@Entity
@Table(name = "survey")
public class Survey {

    @Id
    private Long id;

    @Column
    private String expectation;

    @Column
    private String payFor;

    @Column
    private String need;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getExpectation() {
        return expectation;
    }

    public void setExpectation(String expectation) {
        this.expectation = expectation;
    }

    public String getPayFor() {
        return payFor;
    }

    public void setPayFor(String payFor) {
        this.payFor = payFor;
    }

    public String getNeed() {
        return need;
    }

    public void setNeed(String need) {
        this.need = need;
    }
}
